application-appstore: support for ApplicationScrollSignal method
When the ApplicationScrollSignal is called from indicator-application
the XAyatanaScrollAction method provided by libappindicator is
called with the scroll event data.

So: indicator-application now is listening for that signal
("scroll-entry"), registering a call-back function. When it gets
called, it finds the scrolled application and informs (via dbus)
indicator-application-service that a scroll event has been
registered for that application. Now the service (that is connected
to every indicatorapp running) sends (via dbus, again) to the correct
application the scroll-event informations.