Don't use "SecondaryActivate" method
Using XAyatanaSecondaryActivate as libindicator doesn't support
the mouse x,y position in secondary_activate signal anymore.

This will drop the middle-click support also for KDE status
notifier items, but this is needed as we can't control
what they would do with this signal (and according to our
policies we can't do anything that isn't doable also using
a menu item).